Move machine name resolution out of pkg/agent into internal/provision
Addresses review feedback: pkg/agent is a shared library with external
consumers, so baking AGENT_MACHINE_NAME / hostname resolution into it (and
changing the MachineName JSON tag) risked surprising those consumers.

Relocate the resolution to internal/provision, the unbounded-specific config
layer that already wraps pkg/agent/config and is imported by both the agent
start path and the daemon. pkg/agent is now byte-for-byte unchanged vs main
(no omitempty tag change, no BackfillMachineName, no defensive call in
goalstates.ResolveMachine).

cmd/agent now calls provision.ResolveMachineName before BackfillNodeName at
both AgentConfig entry points (normalizeConfig and FindActiveMachine), which
preserves the guarantee that MachineName is populated before any reader. The
generated config carries an empty "MachineName" (resolved on the host) rather
than omitting the field.
