+++ internal/provision/agent_config_test.go
@@ -119,8 +119,7 @@ func TestAgentConfig_EmptyFields(t *testing.T) {
 	var parsed map[string]interface{}
 	require.NoError(t, json.Unmarshal(data, &parsed))
 
-	// MachineName has omitempty so should be absent from a zero-value config.
-	require.NotContains(t, parsed, "MachineName")
+	require.Equal(t, "", parsed["MachineName"])
 
 	cluster := parsed["Cluster"].(map[string]interface{})
 	require.Equal(t, "", cluster["CaCertBase64"])
